requirements.txt
dash==2.7.0
dash_bio==1.0.1
joblib==1.3.2
numpy==1.24.4
pandas==2.0.3
plotly==5.16.1
rdkit==2023.3.2
scikit_learn==1.3.0
umap_learn==0.5.4
